Fix #15 - split navigator from footer and add to header
